requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.2
django-filter==23.4
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0
python-dateutil==2.8.2
pytz==2023.3
psycopg2-binary==2.9.9